fix(core): disable debug mode for minified build
Debug variable set to true only for dev build in project.clj, but the var was missing in the referenced namespace. Add
the debug var to the core namespace and use it to toggle kee-frame debug mode.
Fix react sequence key issue, moving the key from the anchor tag to the parent.

Fixes #51