feature: plots interaction button improvements
The links "Create new time series", "Create new XY plot" and the icon for saving plots configuration are now presented as proper buttons in a row below the plots. The save plots config button is disabled by default and is enabled if changes are applied to the plots. When saving successfully the button is again disabled.

Closes #145